refactor: bungae -> lightening
