+++ src/api/meeting/index.ts
@@ -1,7 +1,7 @@
 import { ApplicationStatusType, UserResponse } from '@api/user';
 import { api } from '..';
 import { APPROVAL_STATUS_KOREAN_TO_ENGLISH } from '@constants/option';
-import { BungaeFormType } from '@type/form';
+import { LighteningFormType } from '@type/form';
 interface PaginationType {
   page: number;
   take: number;
@@ -47,37 +47,4 @@ export const getMeetingPeopleList = async ({ id, ...rest }: OptionData): Promise
       },
     })
   ).data;
-};
-
-export const createBungae = async (formData: BungaeFormType) => {
-  const {
-    data: { lightningId },
-  } = await api.post<{ lightningId: number }>('/lightning/v2', filterBungaeFormData(formData));
-  return lightningId;
-};
-
-const filterBungaeFormData = (formData: BungaeFormType) => {
-  const convertedTags = formData.welcomeTags?.map(tag => {
-    return tag?.value;
-  });
-  const convertedEndDate =
-    formData.timeInfo.time.value === '당일' ? formData.timeInfo.startDate : formData.timeInfo.endDate;
-  const convertedLightningPlace =
-    formData.placeInfo.place.value === '협의 후 결정' ? null : formData.placeInfo.placeDetail;
-  const data = {
-    lightningBody: {
-      title: formData.title,
-      desc: formData.desc,
-      lightningTimingType: formData.timeInfo.time.value,
-      activityStartDate: formData.timeInfo.startDate,
-      activityEndDate: convertedEndDate,
-      lightningPlaceType: formData.placeInfo.place.value,
-      lightningPlace: convertedLightningPlace,
-      minimumCapacity: formData.minCapacity,
-      maximumCapacity: formData.maxCapacity,
-      files: formData.files,
-    },
-    welcomeMessageTypes: convertedTags?.length === 0 ? null : convertedTags,
-  };
-  return data;
 };
